fix ci sql qta bug

+++ test/send_proposal_request_ci/stop_sql_tbft_4.sh
@@ -4,12 +4,15 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 #
-
 pid=`ps -ef | grep chainmaker | grep "ci-sql-tbft" | grep -v grep |  awk  '{print $2}'`
-for p in $pid
-do
-    kill $p
-    echo "kill $p"
-done
+if test -z "$pid";  then
+  echo "no process to kill"
+else
+  for p in $pid
+  do
+      kill $p
+      echo "kill $p"
+  done
+  ps -ef|grep chainmaker | grep "ci-sql-tbft"
+fi
 
-ps -ef|grep chainmaker | grep "ci-sql-tbft"
